feat: Refactor notification data structure
This commit refactors the notification data structure by:

-   Moving `NotificationDto`, `NotificationModel` and `NotificationRepository` to a new package called `notification`.
-  Moving `NotificationDataSource` to the `notification` package.
- Creating  `NotificationModel` that represents notification information, including id, type, title, description, image URL and timestamp.
- Creating `NotificationDto` to hold raw notification data from API.
- Creating `NotificationRepository` to fetch notification data from the network and handling any potential errors.
- Creating `NotificationDataSource` to fetch notification from the API.
