--- app/src/main/java/com/surivalcoding/composerecipeapp/data/data_source/NotificationDataSource.kt
@@ -1,5 +1,0 @@
-
-
-interface IngredientDataSource {
-    suspend fun getIngredients(): List<IngredientModel>
-}
