+++ app/src/main/java/com/surivalcoding/composerecipeapp/data/recipe/data_source/NotificationDataSource.kt
@@ -1,0 +1,6 @@
+import com.surivalcoding.composerecipeapp.data.recipe.ingrident.IngredientModel
+
+
+interface NotificationDataSource{
+    suspend fun getNotifications(): List<NotificationModel>
+}
